[opentitantool] Add optional --gpio_strap to transport init
Add the ability to specify a named gpio strap to the `transport init`
call. If a strap is specified, then the values from the gpio strap will
be used into of the default pin configuration. This is useful when one
wished to avoid gpio glitches between `transport init` and `gpio apply
<STRAP_NAME>`.

Change-Id: Ic1b0ac44a7e55f1ec7055ab2d6407eae6653c923